requirements.txt
# Core dependencies
anthropic>=0.40.0
python-docx>=1.1.0

# Optional: for local PDF text extraction (backup)
pdfplumber>=0.11.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0  # Beautiful CLI output
click>=8.1.0  # CLI framework
